Renaming components of paragraph menu
